♻️  컴포넌트 분리&네이밍
